refactor: move matcher to top-level
