+++ src/event/synchronize.ts
@@ -1,7 +1,7 @@
 import type { EventTypesPayload, WebhookEvent } from "@octokit/webhooks";
 import type { Context } from "probot";
 
-import { match, PullRequestBase } from "../pr/matcher";
+import { match, PullRequestBase } from "../matcher";
 import { enqueue } from "../tasks/queue";
 import { synchronizeManaged } from "../tasks/synchronize-managed";
 import { synchronizeReview } from "../tasks/synchronize-review";
