+++ src/event/unlabeled.ts
@@ -4,7 +4,7 @@ import type { Context } from "probot";
 import { getConfig } from "../config";
 import { closePullRequest } from "../github/close";
 import { deleteBranch } from "../github/git";
-import { createManaged, PullRequestBase } from "../pr/matcher";
+import { createManaged, PullRequestBase } from "../matcher";
 
 export async function onLabelRemoved(
   context: EventTypesPayload["pull_request.unlabeled"] &
