+++ src/github/wait-for-update.ts
@@ -1,7 +1,7 @@
 import { Context } from "probot";
 import { promisify } from "util";
 
-import { PullRequestBase } from "../pr/matcher";
+import { PullRequestBase } from "../matcher";
 
 const wait = promisify(setTimeout);
